Rename a class with a terribly misleading name
